+++ app/mcts_run_setup.py
@@ -13,6 +13,7 @@ from rostok.criterion.simulation_flags import (EventContact,
 from rostok.simulation_chrono.simulation_scenario import GraspScenario
 from rostok.trajectory_optimizer.control_optimizer import (
     CalculatorWithOptimizationDirect, TendonOptimizerCombinationForce)
+from rostok.utils.numeric_utils import Offset
 
 
 def config_independent_torque(grasp_object_blueprint):
@@ -63,20 +64,6 @@ def config_independent_torque(grasp_object_blueprint):
                                                          hp.CONTROL_OPTIMIZATION_BOUNDS,
                                                          hp.CONTROL_OPTIMIZATION_ITERATION)
     return control_optimizer
-
-
-class Offset:
-
-    def __init__(self, value: float, is_ratio: bool, x_shift: bool = False):
-        self.value = value
-        self.is_ratio = is_ratio
-        self.x_shift = x_shift
-
-    def get_offset(self, x):
-        if self.is_ratio:
-            return x * self.x_shift + x * self.value
-        else:
-            return x * self.x_shift + self.value
 
 
 def config_with_tendon(grasp_object_blueprint):
